+++ TSOClient/tso.common/FSOEnvironment.cs
@@ -21,7 +21,7 @@ namespace FSO.Common
         /// </summary>
         public static bool SoftwareDepth = false;
         public static float UIZoomFactor = 1f;
-        public static int DPIScaleFactor = 1;
+        public static float DPIScaleFactor = 1;
         public static bool SoftwareKeyboard = false;
         public static int RefreshRate = 60;
